+++ src/components/Challenge/common/index.ts
@@ -1,1 +1,4 @@
 export * from './Section';
+export * from './IntroduceCard';
+export * from './ChallengeImage';
+export * from './NumberOfPersion';
